+++ CHANGELOG.md
@@ -6,13 +6,14 @@ Not yet released.
 
 ### Features
 
-
+* [CS-542][] Add `"container-user"` to Docker Server settings. This allows you to specify the user within the container as whom the command should be run. If you specify nothing, the process within the container is run as root or whatever user settings are defined on the image (which is unchanged from the current behavior).
 
 ### Bugfixes
 
 * [CS-541][] Use Path Translation setting when creating mounts for setup and wrapup commands.
 
 [CS-541]: https://issues.xnat.org/browse/CS-541
+[CS-542]: https://issues.xnat.org/browse/CS-542
 
 ## 1.6.0